refactor(cli): extract arg parsing + recipe overlay to cli/args.ts
